pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-zotero"
version = "0.1.0"
description = "MCP server (Python) untuk proyek mcp-zotero"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pypdf>=5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
]

[project.scripts]
mcp-zotero = "mcp_zotero.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
